Refactor: Move monitor package from core to seedNode and rename to reporting
